fix: #93 --upgrade-type ignored
If `--requirements` or `--project` flags are used, current_version for
Packages was equal to the specifier in the requirements file
eg. requests==2.24.0 showed current_version as `==2.24.0` rather than
just `2.24.0`
This then caused the `upgrade_type` check later to fail (as the `==`
version prefix made the version difference calculation incorrect

Requirements will now include the version only if an exact version is
specified, otherwise the installed version is used.
If not installed, the entire specifier is used (eg. `~=1.0`)